Prokudin-Gorsky: registration by maximization of mutual information
This is an implementation and demonstration of registration by maximization of
mutual information. A method to assign photographic plates to corresponding
color channels is also included.

This computer vision algorithm is demonstrated on an input of historical note
for photography, the color plates of Prokudin-Gorsky.

I developed this project in Professor Learned-Miller's graduate computer vision
class.